testing 123
In this exercise, we'll work with basic Python functions, which will involve some NumPy. This in no way is suppose to be an introduction to Python and this notebook will assume familiarity with the language. It's purpose in the end is to brush up the user and to serve as a "mini-prerequisite" for the rest of the notebook. Next, we'll play around with an image in order to get some exposure towards working with pixels and matrices. In the process we'll also show how your eyes are more sensitive towards different types of color and shading. Lastly, we'll create a homemade kNN model and test it against sklearn's model.

The following problems have taken heavy inspiration from UC Berkeley's DS100 class (https://github.com/DS-100). The purpose of this notebook is to practice creating "teaching" notebooks and to further improve my data science skills by solving these problems. Aside from the problems, most of the work (questions, explainations, solutions, and creation of notebook) are mine.